fix: change pointerupoutside event callback #2674 #2659
