+++ common/changes/@visactor/vtable/fix-pointerupoutside_2024-10-28-12-12.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@visactor/vtable",
+      "comment": "fix: change pointerupoutside event callback #2674 #2659",
+      "type": "none"
+    }
+  ],
+  "packageName": "@visactor/vtable"
+}
